refactor: move Contents directory to components
